+++ src/com/company/Model/IntrinsicValue.java
@@ -7,7 +7,7 @@ public class IntrinsicValue {
     private Double cashTakenOutOfBusiness; // coupon -> Cash Taken out of business -> This is dividends recieved for 1 year
     private Double currentBookValue; // par -> Current Book Value -> We need to know this so we can determine the base value that's changing
     private Double averagePercentChangeInBookValuePerYear; // bvc -> Average Percent Change in book value per year -> This will determine the estimate BV at the end of the next 10 years
-    private Integer years; // year -> Years -> This will most likely be 10 (if you're comparing a 10 year federal note)
+    private Double years; // year -> Years -> This will most likely be 10 (if you're comparing a 10 year federal note)
     private Double tenYearFederalNote; // r -> 10 year federal note -> Look up the ten year treasury note by clicking on this text
 
     public Double getCashTakenOutOfBusiness() {
@@ -34,11 +34,11 @@ public class IntrinsicValue {
         this.averagePercentChangeInBookValuePerYear = averagePercentChangeInBookValuePerYear;
     }
 
-    public Integer getYears() {
+    public Double getYears() {
         return years;
     }
 
-    public void setYears(Integer years) {
+    public void setYears(Double years) {
         this.years = years;
     }
